Keep the fraction in a merged peak, as the reference does
deduplicate_peaks in the published postprocessor replaces a group of adjacent peak frames with their mean, and that mean is a real number; it becomes a time only at the end. This port accumulated it in an int, so integer division dropped the fraction at every step and a plateau of frames 10, 11 and 12 came out at 10 rather than 10.5 - a systematic half-frame shift on every merged beat, 10ms at this frame rate.

Peaks and RawPeaks now return fractional frames, Prune and LocalPeriod take them, and reading the activation a peak was found in goes through FrameOf. The new test carries the reference algorithm transcribed and compares it with what the tracker produces, including the quirk that follows from comparing against a running mean: the third frame of a three-frame plateau starts a group of its own.

Measured on the reference track with the float model against the beatmap's own grid: 551 beats, median residual 29ms, p90 137.7ms, 34.8% more than 60ms out - the same median, a slightly better p90 and a slightly better spread than before. The coarser metrical-level count moves from 2 to 3 points of 21 because the tracker now sits on the other level the map declares in more places, which is the level its own policy says to keep.

The README now states the pruning pass as the deliberate difference from the reference that it is, next to the resampler.

+++ ParaTactus/StreamingBeatTracker.cs
@@ -198,7 +198,7 @@ namespace ParaTactus
 
             var peaks = BeatThisBeatTracker.Peaks(activation.ToArray());
 
-            foreach (int peak in peaks)
+            foreach (double peak in peaks)
             {
                 if (peak < emittedThrough || peak >= limit)
                     continue;
